Add support for processor data
This update adds support for storing a state to be
used by parsers, compilers, and plug-ins, by exposing
an optional `data` property on `processor`, `Processor`,
and `Processor#`.

This property holds a `JSON.stringify`able object, which
should be given to `unified` in its options. Then, that
value is exposed as `Processor.data` and `Processor#data`.

Later, when constructing a processor, that value is
cloned and exposed as `processor.data`.

This change additionally ensures `processor` is given to
`Parser` and `Compiler` when constructing as a third
parameter.

Related to wooorm/hast#4.