Remove `processor` as first param from attachers
MAJOR BREAKING CHANGE: The first parameters is now no longer given
as an argument to attachers (a.k.a., plug-ins).  The processor is
now given to attachers as the context object (“this”) instead.

This now allows plugins to be more easily used without unified
(something that I believe is good for the ecosystem).  Additionally,
this makes plugins usable inside other plugins as “utilities”, which
was previously rather awkward: https://github.com/wooorm/rehype-format/
blob/7715bf1/index.js#L4

Closes GH-21.