Set `file.result` when processing to non-text
unified is typically, but not always, used to *serialize* when it
compiles.  Compilers don’t always return such a serialized (`string`,
`Buffer`) though.  The previous behavior was to place the result of
the stringify phase when processing on `file.contents`.  This doesn’t
make sense for non-text.  This change places non-text results on
`file.result`.

Closes vfile/vfile#45.