types!: allow pluggable list to check settings on first 20 elements
leverages variadic typing and generics to allow the first 20 plugins to have full settings type checking.
plugins after the first twenty can still be added, but no longer have their settings checked.

BREAKING CHANGE: changes to order of the processor settings generic `P`, to go first and plugin settings `S` to go after. Plugins directly using the `Plugin` or `Attacher` type will need to add the first generic with the processor settings.
NOTE: If the plugin does not directly use the `Plugin` or `Attacher` type it may not need to migrate. For example
```ts
// this does NOT need to migrate, it will continue to work
const plugin = () => (settings: Settings) => { /* ... */ }

// this DOES need to migrate, settings would passed to the wrong generic parameter
const plugin: Plugin<[Settings?]> = () => (settings: Settings) => { /* ... */ }
```
MIGRATION: if `this` is not used in the plugin, set the first generic to `unknown`
For example
```typescript
Plugin<[PluginSettings?]>
// becomes
Plugin<unknown, [PluginSettings?]>

// and

Attacher<[PluginSettings?]>
// becomes
Attacher<unknown, [PluginSettings?]>
```
if it does depend on `this`, add the processor options in the first generic, for example `RemarkOptions` or `RehypeOptions` could be used.