Fix to set `vfile.value` if a vfile@5 is given
vfile@5 changed from the `contents` to the `value field.
Given that not all of the ecosystem has updated yet, and that some
version mismatches will occur for the foreseeable future, setting
both `contents` *and* `value` in those cases will hopefully alleviate
some of the struggles with updating.

Related to remarkjs/remark-frontmatter#16.